Refactor the project to adhere to https://github.com/golang-standards/project-layout
