+++ Makefile
@@ -1,6 +1,6 @@
 .PHONY: all clean install uninstall
 
-all: popub-local/popub-local popub-relay/popub-relay
+all: popub-local popub-relay
 
 PREFIX=/usr/local
 GOBUILD=go build
@@ -9,18 +9,17 @@ clean:
 	rm -f popub-local/popub-local popub-relay/popub-relay
 
 install: all
-	install -Dm0755 popub-local/popub-local "$(DESTDIR)$(PREFIX)/bin/popub-local"
-	install -Dm0755 popub-relay/popub-relay "$(DESTDIR)$(PREFIX)/bin/popub-relay"
+	install -Dm0755 popub-local "$(DESTDIR)$(PREFIX)/bin/popub-local"
+	install -Dm0755 popub-relay "$(DESTDIR)$(PREFIX)/bin/popub-relay"
 	$(MAKE) -C systemd install DESTDIR="$(DESTDIR)" PREFIX="$(PREFIX)"
 
 uninstall:
 	rm -f "$(PREFIX)/bin/popub-local" "$(DESTDIR)$(PREFIX)/bin/popub-relay"
 	$(MAKE) -C systemd uninstall DESTDIR="$(DESTDIR)" PREFIX="$(PREFIX)"
 
-popub-local/popub-local: popub-local/main.go popub-local/delayer.go
-	cd popub-local && $(GOBUILD)
+popub-local:
+	$(GOBUILD) github.com/m13253/popub/cmd/popub-local
 
-popub-relay/popub-relay: popub-local/main.go popub-local/delayer.go
-	cd popub-relay && $(GOBUILD)
-
+popub-relay:
+	$(GOBUILD) github.com/m13253/popub/cmd/popub-relay
