+++ README.md
@@ -53,7 +53,7 @@ Note: The password is for authorization only, your traffic is not encrypted.
 Running as Systemd services
 ---------------------------
 
-Refer to [systemd/README.md](systemd/README.md) for instructions on running as Systemd services.
+Refer to [systemd/README.md](contrib/systemd/README.md) for instructions on running as Systemd services.
 
 License
 -------
